// src/app/data/models/became-creator.ts
export interface BecameCreator {
  id?:                       number;
  first_name:               string;
  last_name:                string;
  email:                    string;
  phone?:                   string;
  principal_social_network: string;
  pseudo_social_network:    string;
  followers:                string;
  status:                   string;
  date?:                     string;
}